Refactor project structure and update paths
Moved files from the base `pages/` and `scripts/` directories into the `src/` directory for better organization. Updated file paths in HTML, CSS, scripts, and test files to reflect the new structure. Adjusted `.gitignore` to include `package-lock.json` and refined asset paths in stylesheets.
