+++ .github/workflows/static.yml
@@ -35,7 +35,7 @@ jobs:
           node-version: '22'
 
       - name: Install Dependencies
-        run: npm ci
+        run: npm install
 
       - name: Build Project
         run: npm run build
