+++ .gitignore
@@ -3,3 +3,5 @@ node_modules/
 coverage/
 .env/
 dist/
+
+package-lock.json
