+++ index.html
@@ -14,7 +14,7 @@
 
     <meta charset="UTF-8">
     <title>Landing?</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <link rel="stylesheet" href="src/styles/main.css">
 </head>
 <body>
 <!-- Google Tag Manager (noscript) -->
@@ -48,15 +48,15 @@
     <!-- footer: links to various parts of site, copyright, messaging feature? -->
     <footer class="footer"></footer>
 
-    <div id="data-files" data-source-pages="pages/" data-source-assets="./assets/"></div>
-    <script type="text/javascript" src="scripts/data-functions.js"></script>
+    <div id="data-files" data-source-pages="src/pages/" data-source-assets="./assets/"></div>
+    <script type="text/javascript" src="src/scripts/data-functions.js"></script>
 
-    <script type="text/javascript" src="scripts/main.js"></script>
-    <script type="text/javascript" src="scripts/date-time-functions.js"></script>
-    <script type="text/javascript" src="scripts/subpage-init.js"></script>
-    <script type="module" src="scripts/subpage-manipulation.js"></script>
-    <script type="module" src="scripts/posts-functions.js"></script>
-    <script type="module" src="scripts/listeners.js"></script>
+    <script type="text/javascript" src="src/scripts/main.js"></script>
+    <script type="text/javascript" src="src/scripts/date-time-functions.js"></script>
+    <script type="text/javascript" src="src/scripts/subpage-init.js"></script>
+    <script type="module" src="src/scripts/subpage-manipulation.js"></script>
+    <script type="module" src="src/scripts/posts-functions.js"></script>
+    <script type="module" src="src/scripts/listeners.js"></script>
     <!--suppress JSUnresolvedLibraryURL -->
     <script src="https://kit.fontawesome.com/a7ada04e46.js" crossorigin="anonymous"></script>
 </body>
