+++ public/table-of-contents.html
@@ -1,6 +1,5 @@
-
 <!DOCTYPE html>
-<html lang="en-us">
+<html lang="en">
 <head>
     <!-- Google Tag Manager -->
     <!--suppress EqualityComparisonWithCoercionJS -->
@@ -12,14 +11,15 @@
     <!-- End Google Tag Manager -->
 
     <meta charset="UTF-8">
-    <title>Project Page</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <title>Directory</title>
+    <link rel="stylesheet" href="../src/styles/main.css">
 </head>
 <body>
 <!-- Google Tag Manager (noscript) -->
 <noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-5XJHCP3G"
                   height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
 <!-- End Google Tag Manager (noscript) -->
+<!-- javascript warning for any javascript replaced sections -->
 
 <section class="menu"></section>
 
@@ -28,30 +28,27 @@
 
 <div class="banner-small"></div>
 
-<main class="row gtr-0">
-    <!--  -->
-    <div class="col-8 col-12-narrower style-featured-2">
-        <div class="inner">
-            <div class="major"><h2>Cenetral</h2></div>
+<main>
+    <div class="inner row">
+        <div class="col-12">
+            <h2>Directory</h2>
             <div class="table"></div>
         </div>
     </div>
-    <div class="col-4 col-12-narrower"><div class="featured-posts-sidebar"></div></div>
 </main>
-
-<div class="landing-links"></div>
 
 <footer class="footer"></footer>
 
-<div id="data-files" data-source-pages="pages/" data-source-assets="./assets/"></div>
-<script type="text/javascript" src="scripts/data-functions.js"></script>
-<script type="text/javascript" src="scripts/main.js"></script>
-<script type="text/javascript" src="scripts/date-time-functions.js"></script>
-<script type="text/javascript" src="scripts/subpage-init.js"></script>
-<script type="module" src="scripts/subpage-manipulation.js"></script>
-<script type="module" src="scripts/posts-functions.js"></script>
-<script type="module" src="scripts/listeners.js"></script>
+<div id="data-files" data-source-pages="src/pages/" data-source-assets="./assets/"></div>
+<script type="text/javascript" src="../src/scripts/data-functions.js"></script>
+<script type="text/javascript" src="../src/scripts/main.js"></script>
+<script type="text/javascript" src="../src/scripts/date-time-functions.js"></script>
+<script type="text/javascript" src="../src/scripts/subpage-init.js"></script>
+<script type="module" src="../src/scripts/subpage-manipulation.js"></script>
+<script type="module" src="../src/scripts/posts-functions.js"></script>
+<script type="module" src="../src/scripts/listeners.js"></script>
 <!--suppress JSUnresolvedLibraryURL -->
 <script src="https://kit.fontawesome.com/a7ada04e46.js" crossorigin="anonymous"></script>
+
 </body>
 </html>
